Made Login it's own component, rendered within homepage, playlists, and subscriptions. Please run locally and ensure each three pages let you login (refresh per page so you can test logging into each). Note, the subscriptions page does not show anything, and the homepage and playlist page render practically the same thing (for now). On success, it should display a welcome message and your playlists (on homepage and playlists page).
