+++ src/pages/Playlists.tsx
@@ -1,23 +1,16 @@
 import { useAppSelector } from "../redux/hooks";
-import { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
 import AllPlaylists from "../components/AllPlaylists";
 import SinglePlaylist from "../components/SinglePlaylist";
 
+import Login from "../components/Login";
+
 function Playlists() {
-  const navigate = useNavigate();
   const playlists = useAppSelector((state) => state.playlists.playlists);
   const user = useAppSelector((state) => state.user.info);
 
-  // redirect to root page if user is not authenticated
-  useEffect(() => {
-    if (!user?.access_token) {
-      navigate("/");
-    }
-  }, [user, navigate]);
-
   return (
     <>
+      <Login />
       <h1>
         {user?.given_name ? `${user.given_name}'s Playlists` : "Playlists"}
       </h1>
